requirements.txt
einops
ema_pytorch
ftfy
fvcore
git+https://github.com/openai/CLIP.git
h5py
iopath
regex
scipy
torch==2.4.1
torchvision==0.19.1
tqdm==4.66.5
transformers==4.45.2
numpy==1.26.4
scikit-learn==1.5.2
evo==1.30.0
hydra-core==1.3.2
torchtyping==0.1.5
lightning==2.4.0
lightning[extra]
triton==3.0.0
python-dotenv==1.0.1
